fix(integer-overflow): type-width modeling, unsigned-sub coverage, definite tier
- IO-02/16/17: model integer width. A 64-bit operand (size_t/long/sizeof)
  promotes the product to 64 bits on LP64, so `n * sizeof(T)` / `size_t * x`
  no longer emit a CWE-190 candidate; 32-bit (`int * int`, `int * 1024`)
  products still do. Sync benchmark P8-01 and the tc30/tc64/tc66/tc93 fixtures.
- IO-07: an influenced local (a parameter, or a local assigned from a call/
  influenced variable) now gates the mul-const pattern, not just a parameter.
- IO-08: unsigned-sub underflow now also scans assignment RHS
  (`size_t rem = total - consumed;`) and uses the influenced set.
- IO-09: detect arithmetic by the AST operator token, so `n*m` without
  whitespace is recognized.
- IO-10: resolve multi-level assignment chains (`t = n*m; u = t; malloc(u)`).
- IO-25: a pure-literal product that provably overflows int32 emits
  `definite_overflow` and is auto-confirmed.

+++ sgre/internal/planner/filter_int_overflow_test.go
@@ -92,14 +92,14 @@ func TestIntOverflowGuardFilter_MulConst(t *testing.T) {
 	path := filepath.Join(dir, "iof_add.c")
 	src := `#include <stdlib.h>
 
-void guarded_mul(size_t n) {
+void guarded_mul(int n) {
     if (n < 100) {
         char *buf = malloc(n * 1024);
         (void)buf;
     }
 }
 
-void unguarded_mul(size_t n) {
+void unguarded_mul(int n) {
     char *buf = malloc(n * 1024);
     (void)buf;
 }
